refactor(settings): move SettingsGeneral and SettingsYou to widgets directory
